Avoid problems on 32-bit systems (#1159)
* Apply changes from #1156 to v2

* Avoid problems on 32-bit systems

+++ jwk/jwk_test.go
@@ -1503,7 +1503,7 @@ c4wOvhbalcX0FqTM3mXCgMFRbibquhwdxbU=
 
 type typedField struct {
 	Foo string
-	Bar int
+	Bar int64
 }
 
 func TestTypedFields(t *testing.T) {
